W112FclCombEff: reconstruct Fcl effect paths and suffix float constants (18c70 616/992->776/992 nd429->557; 19f20 700/928->728/928 nd532->546; 1aff0 1068/1344->924/1344 nd722->639)

+++ src/Event/Fcl/fclShopMisc.c
@@ -4297,7 +4297,7 @@ u32 FUN_003f3f00(u64 param_1)
 // FUN_003F3FC0 NONMATCHING
 
 
-void FUN_003f3fc0(int param_1,int param_2,u32 param_3,int param_4)
+void FUN_003f3fc0(int param_1,int param_2,u32 param_3,u64 param_4)
